coordinator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0
slowapi>=0.1.9
structlog>=24.1.0